requirements.txt
pyspark
langchain
openai
faiss-cpu
tiktoken
pandas
numpy
scikit-learn
fastapi
uvicorn
python-dotenv
pyyaml
sentence-transformers
dspy-ai
